requirements.txt
openpyxl~=3.1.5
python-docx~=1.1.2
lxml~=5.3.0
flask
werkzeug